Users and permissions

+++ fars/booking/views.py
@@ -7,13 +7,19 @@ from datetime import datetime, timedelta
 
 def home(request):
     bookables = Bookable.objects.all()
-    context = {'bookables': bookables}
+    context = {
+        'bookables': bookables,
+        'user': request.user,
+    }
     return render(request, 'base.html', context)
 
 
 def bookings_month(request, bookable):
     bookable_obj = get_object_or_404(Bookable, id_str=bookable)
-    context = {'bookable': bookable_obj}
+    context = {
+        'bookable': bookable_obj,
+        'user': request.user
+    }
     return render(request, 'month.html', context)
 
 
@@ -22,18 +28,27 @@ def bookings_day(request, bookable, year, month, day):
     context = {
         'date': "{y}-{m:02d}-{d:02d}".format(y=year, m=month, d=day),
         'bookable': bookable_obj,
+        'user': request.user
     }
     return render(request, 'day.html', context)
 
 
 def book(request, bookable):
+    if not request.user.is_authenticated:
+        return render(request, 'modals/forbidden.html')
     booking = Booking()
     bookable_obj = get_object_or_404(Bookable, id_str=bookable)
-    context = {'url': request.path, 'bookable': bookable_obj}
+    context = {
+        'url': request.path,
+        'bookable': bookable_obj,
+        'user': request.user,
+    }
+
     if request.method == 'GET':
         start = datetime.strptime(request.GET['t'], '%Y-%m-%dT%H:%M:%S') \
             if 't' in request.GET else datetime.now()
         booking.bookable = bookable_obj
+        booking.user = request.user
         booking.start = start
         booking.end = start + timedelta(hours=1)
         form = BookingForm(instance=booking)
@@ -53,11 +68,16 @@ def book(request, bookable):
 
 def unbook(request, booking_id):
     booking = get_object_or_404(Booking, id=booking_id)
-    if request.method == 'POST':
+    # TODO: is this the proper way to check if users are the same?
+    unbookable = (request.user.username == booking.user.username) \
+        or request.user.is_staff
+    if request.method == 'POST' and unbookable:
         booking.delete()
         return HttpResponse()
     context = {
         'url': request.path,
         'booking': booking,
+        'user': request.user,
+        'unbookable': unbookable,
     }
     return render(request, 'unbook.html', context)
